Refactor: move `calculateTimeLeft` to utils, update `CountdownTimer`
